[DataAnalytics] 테이블 정보 우측 하측 trim
