+++ src/assets/entry/scss/popup.scss
@@ -2475,8 +2475,8 @@
     flex: 1;
     position: relative;
     overflow: hidden;
+    margin-left: -27px;
     .inner {
-        margin-left: -27px;
         position: relative;
         -webkit-box-flex: 1;
         flex: 1 1;
