+++ src/components/editor/dataAnalytics/DataAnalyticsEditor.jsx
@@ -1,6 +1,4 @@
 import React, { useContext } from 'react';
-import ReactDOM from 'react-dom';
-import Header from './Header';
 import Summary from './summary/Summary';
 import TableEditor from './TableEditor';
 import ChartEditor from './chart/ChartEditor';
@@ -64,89 +62,3 @@ const DataAnalyticsEditor = (props) => {
 };
 
 export default DataAnalyticsEditor;
-
-// const chart = (params) => {
-//     return (
-//         <div className={Styles.chart_box}>
-//             <h2 className={Styles.blind}>차트</h2>
-//             <div className={Styles.inner}>
-//                 {/* 테이블 선택 */}
-//                 <div className={Styles.chart_navi}>
-//                     <a className={`${Styles.chart_link} ${Styles.bar} ${Styles.disabled}`}>
-//                         <span className={Styles.blind}>막대형</span>
-//                     </a>
-//                     <a className={`${Styles.chart_link} ${Styles.line} ${Styles.disabled}`}>
-//                         <span className={Styles.blind}>꺽은선형</span>
-//                     </a>
-//                     <a className={`${Styles.chart_link} ${Styles.pie} ${Styles.disabled}`}>
-//                         <span className={Styles.blind}>원형</span>
-//                     </a>
-//                     <a className={`${Styles.chart_link} ${Styles.scatter} ${Styles.disabled}`}>
-//                         <span className={Styles.blind}>방사형</span>
-//                     </a>
-//                     <div className={Styles.add_link_box}>
-//                         <a className={Styles.add_link} role="button">
-//                             <span className={Styles.blind}>추가하기</span>
-//                         </a>
-//                     </div>
-//                 </div>
-//                 {/* 테이블 차트 입력폼 */}
-//                 <div className={Styles.chart_no_result}>
-//                     <p className={Styles.dsc}>차트를 먼저 추가해 주세요.</p>
-//                 </div>
-//             </div>
-//         </div>
-//     );
-// };
-
-// const Portal = ({ children }) => {
-//     const el = document.querySelector('body');
-//     return ReactDOM.createPortal(children, el);
-// };
-// const DataAnalyticsEditor2 = () =
-// {
-//     const { dataAnalytics, dispatch } = useContext(DataAnalyticsContext);
-//     const { tab, table, isFullScreen } = dataAnalytics;
-
-//     const handleFullScreenClick = (event) => {
-//         event.preventDefault();
-//         dispatch({
-//             type: 'TOGGLE_FULLSCREEN',
-//             isFullScreen: !isFullScreen,
-//         });
-//     };
-
-//     const header = (
-//         <Header
-//             selected={tab}
-//             tabItems={TAB_ITEMS}
-//             isFullScreen={isFullScreen}
-//             onFullScreenClick={handleFullScreenClick}
-//         />
-//     );
-//     let content = null;
-//     if (table) {
-//         switch (tab) {
-//             case SUMMARY:
-//                 content = <Summary />;
-//                 break;
-//             case TABLE:
-//                 content = <TableEditor />;
-//                 break;
-//             case CHART:
-//                 content = <ChartEditor />;
-//                 break;
-//             default:
-//                 break;
-//         }
-//     }
-
-//     return (
-//         <Portal>
-//             <div className={`${Styles.data_detail_wrap} ${Styles.full}`}>
-//                 {header}
-//                 {content}
-//             </div>
-//         </Portal>
-//     );
-// };
