+++ src/components/editor/dataAnalytics/chart/HorizontalLegend.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { categoryKeys } from '@utils/Common';
+import { categoryKeys } from '@utils/dataAnalytics';
 import { GRAPH_COLOR } from '@constants/dataAnalytics';
 import Theme from '@utils/Theme';
