+++ src/components/editor/dataAnalytics/chart/YAxis.jsx
@@ -1,7 +1,8 @@
 import React, { useContext, useState, useRef } from 'react';
 import { DataAnalyticsContext } from '@contexts/dataAnalytics';
 import Dropdown from '@components/widget/dropdown';
-import { CommonUtils, getNumberColumnIndexes } from '@utils/Common';
+import { CommonUtils } from '@utils/Common';
+import { getNumberColumnIndexes, getTrimedTable } from '@utils/dataAnalytics';
 import Theme from '@utils/Theme';
 
 const YAxis = () => {
@@ -12,7 +13,7 @@ const YAxis = () => {
     const { selected = {} } = dataAnalytics;
     const { fields = [], origin = [], chart, chartIndex } = selected;
     const { yIndex = 0, xIndex } = chart[chartIndex];
-    const table = [[...fields], ...origin];
+    const table = getTrimedTable([[...fields], ...origin]);
     const disabled = xIndex === -1;
     const yAxisIndex = disabled ? [] : getNumberColumnIndexes(table, [xIndex]);
     const yAxis = yAxisIndex.map((index) => [fields[index], index]);
