+++ src/components/editor/dataAnalytics/summary/Table.jsx
@@ -1,7 +1,8 @@
 import React, { useContext } from 'react';
 import { DataAnalyticsContext } from '@contexts/dataAnalytics';
 import TableToolTip from '@components/tooltip/TableToolTip';
-import { CommonUtils, getSummary } from '@utils/Common';
+import { getSummary, getTrimedTable } from '@utils/dataAnalytics';
+import { CommonUtils } from '@utils/Common';
 import { SUMMARY_HEADER } from '@constants/dataAnalytics';
 import _map from 'lodash/map';
 import Theme from '@utils/Theme';
@@ -11,7 +12,7 @@ const Table = () => {
     const { dataAnalytics } = useContext(DataAnalyticsContext);
     const { selected = {} } = dataAnalytics;
     const { fields = [], origin = [] } = selected;
-    const table = [[...fields], ...origin];
+    const table = getTrimedTable([[...fields], ...origin]);
     const summary = getSummary(table) || [];
 
     return (
