+++ src/utils/Common.js
@@ -3,9 +3,6 @@ import _ from 'lodash';
 import { DEFAULT_OPTIONS } from '../constants/index';
 import get from 'lodash/get';
 import _some from 'lodash/some';
-import flow from 'lodash/fp/flow';
-import map from 'lodash/fp/map';
-import unzip from 'lodash/fp/unzip';
 
 export const CommonUtils = {
     getScaleNumber(num, inMin, inMax, outMin, outMax) {
@@ -257,91 +254,3 @@ export function FormAsyncException(obj) {
         this[key] = obj[key];
     });
 }
-
-export const isString = (str) => isNaN(str) || Number(str).toString() != str;
-export const someString = (array) => _some(array, isString);
-export const getHeader = (matrix, editable = true) =>
-    _.chain(matrix)
-        .head()
-        .map((column) => ({
-            editor: editable ? 'text' : '',
-            name: column,
-            align: 'center',
-            width: 130,
-            ellipsis: true,
-        }))
-        .value();
-export const getData = (matrix) =>
-    _.chain(matrix)
-        .slice(1)
-        .map((content) => _.zipObject(_.head(matrix), content))
-        .value();
-
-export const toFixed = (num, dp = 2) => Math.round(num * Math.pow(10, dp)) / Math.pow(10, dp);
-const getAverage = (array) => array.reduce((sum, value) => sum + Number(value), 0) / array.length;
-const getStandardDeviation = (arr, average) =>
-    Math.sqrt(arr.reduce((acc, curr) => acc + Math.pow(curr - average, 2), 0) / arr.length);
-const makeSummary = (row) => {
-    const restRow = row.slice(1);
-    const count = restRow.length;
-    if (someString(restRow)) {
-        return [row[0], count, '-', '-', '-', '-', '-'];
-    }
-    const max = Math.max(...restRow);
-    const min = Math.min(...restRow);
-    const average = getAverage(restRow);
-
-    return [
-        row[0],
-        average,
-        getStandardDeviation(restRow, average),
-        max,
-        restRow.sort((a, b) => a - b)[Math.floor((restRow.length - 1) / 2)],
-        min,
-    ].map((value) => (isString(value) ? value : toFixed(value)));
-};
-export const getSummary = flow(unzip, map(makeSummary));
-
-export const categoryKeys = (table, index) =>
-    // eslint-disable-next-line prettier/prettier
-    index >= 0 ? _.uniq(table.slice(1).map((row) => row[index])) : [];
-
-export const isZipable = (table, xIndex) =>
-    _.uniqBy(table, (row) => row[xIndex]).length !== table.length;
-
-export const hasNumberColumn = (table) =>
-    _.some(table[0], (columnHeader, columnIndex) =>
-        _.every(table.slice(1), (row) => !isString(row[columnIndex]))
-    );
-
-export const getNumberColumnIndexes = (table, banIndexes = []) =>
-    _.reduce(
-        table[0],
-        // eslint-disable-next-line no-confusing-arrow
-        (prev, curr, index) =>
-            !_.some(banIndexes, (banIndex) => index === banIndex) &&
-            !_.some(table.slice(1), (row) => isString(row[index]))
-                ? [...prev, index]
-                : prev,
-        []
-    );
-
-export const getNumberColumnIndexesBySelectedColumns = (table, selectedColumns = []) =>
-    _.reduce(
-        selectedColumns,
-        // eslint-disable-next-line no-confusing-arrow
-        (prev, index) =>
-            !_.some(table.slice(1), (row) => isString(row[index])) ? [...prev, index] : prev,
-        []
-    );
-
-export const makeTableByGrid = (gridRef) => {
-    const { current = {} } = gridRef;
-    const instance = (current.getInstance && current.getInstance()) || {};
-    const data = (instance.getData && instance.getData()) || {};
-    const columns = (instance.getColumns && instance.getColumns()) || {};
-    return [
-        columns.map((column) => column.name),
-        ...data.map((row) => columns.map((column) => row[column.name])),
-    ];
-};
